+++ Kitodo/src/main/java/org/kitodo/converter/ProcessConverter.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package de.sub.goobi.converter;
+package org.kitodo.converter;
 
 import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
@@ -18,14 +18,13 @@ import javax.faces.convert.ConverterException;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.kitodo.data.database.beans.Ruleset;
+import org.kitodo.data.database.beans.Process;
 import org.kitodo.data.database.exceptions.DAOException;
 import org.kitodo.services.ServiceManager;
 
-public class RegelsatzConverter implements Converter {
-    public static final String CONVERTER_ID = "RegelsatzConverter";
-    private static final Logger logger = LogManager.getLogger(RegelsatzConverter.class);
+public class ProcessConverter implements Converter {
     private final ServiceManager serviceManager = new ServiceManager();
+    private static final Logger logger = LogManager.getLogger(ProcessConverter.class);
 
     @Override
     public Object getAsObject(FacesContext context, UIComponent component, String value) throws ConverterException {
@@ -33,8 +32,8 @@ public class RegelsatzConverter implements Converter {
             return null;
         } else {
             try {
-                return serviceManager.getRulesetService().getById(Integer.valueOf(value));
-            } catch (DAOException | NumberFormatException e) {
+                return serviceManager.getProcessService().getById(Integer.valueOf(value));
+            } catch (NumberFormatException | DAOException e) {
                 logger.error(e);
                 return "0";
             }
@@ -45,12 +44,12 @@ public class RegelsatzConverter implements Converter {
     public String getAsString(FacesContext context, UIComponent component, Object value) throws ConverterException {
         if (value == null) {
             return null;
-        } else if (value instanceof Ruleset) {
-            return String.valueOf(((Ruleset) value).getId().intValue());
+        } else if (value instanceof Process) {
+            return String.valueOf(((Process) value).getId().intValue());
         } else if (value instanceof String) {
             return (String) value;
         } else {
-            throw new ConverterException("Falscher Typ: " + value.getClass() + " muss 'Regelsatz' sein!");
+            throw new ConverterException("Falscher Typ: " + value.getClass() + " muss 'Prozess' sein!");
         }
     }
